requirements.txt
# asgiref==3.5.2
# attrs==22.1.0
# autobahn==22.7.1
# Automat==20.2.0
# certifi==2022.9.24
# cffi==1.15.1
# channels==3.0.5
# charset-normalizer==2.1.1
# constantly==15.1.0
# cryptography==38.0.1
# daphne==3.0.2
# defusedxml==0.7.1
# distlib==0.3.6
# Django==4.0
# django-admin-interface==0.19.2
# django-ckeditor==6.5.1
# django-colorfield==0.7.2
# django-flat-responsive==2.0
# django-flat-theme==1.1.4
# django-froala-editor==4.0.14
# django-js-asset==2.0.0
# django-mathfilters==1.0.0
# djangorestframework==3.13.1
# filelock==3.8.0
# gunicorn==20.1.0
# hyperlink==21.0.0
# idna==3.4
# incremental==21.3.0
# oauthlib==3.2.2
# Pillow==9.2.0
# platformdirs==2.5.2
# psycopg2-binary==2.9.3
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycodestyle==2.9.1
# pycparser==2.21
# PyJWT==2.6.0
# pyOpenSSL==22.1.0
# python3-openid==3.2.0
# pytz==2022.2.1
# requests==2.28.1
# requests-oauthlib==1.3.1
# routing==0.2.0
# service-identity==21.1.0
# six==1.16.0
# social-auth-app-django==5.0.0
# social-auth-core==4.3.0
# sqlparse==0.4.2
# toml==0.10.2
# Twisted==22.8.0
# twisted-iocpsupport==1.0.2
# txaio==22.2.1
# typing_extensions==4.3.0
# tzdata==2022.1
# urllib3==1.26.12
# virtualenv==20.16.5
# websocket-client==1.4.1
# whitenoise==6.2.0
# xlwt==1.3.0
# zope.interface==5.4.0
asgiref==3.5.2
Django==4.0
django-admin-interface==0.19.2
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-froala-editor==4.0.14
django-js-asset==2.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
xlwt==1.3.0
platformdirs==2.5.2
social-auth-app-django==5.0.0
social-auth-core==4.3.0